Bring back /api/v1/vulnerability/ and /api/v1/vulnerability/{identifier}.
