+++ entity/src/vulnerability.rs
@@ -15,8 +15,12 @@ pub struct Model {
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
+    #[sea_orm(has_many = "super::advisory::Entity")]
+    Advisory,
+
     #[sea_orm(has_many = "super::advisory_vulnerability::Entity")]
     AdvisoryVulnerabilities,
+
     #[sea_orm(has_many = "super::vulnerability_description::Entity")]
     Descriptions,
 
@@ -31,6 +35,16 @@ pub enum Relation {
 
     #[sea_orm(has_many = "super::not_affected_package_version::Entity")]
     NotAffectedPackageVersion,
+}
+
+impl Related<advisory::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::Advisory.def()
+    }
+
+    fn via() -> Option<RelationDef> {
+        Some(advisory_vulnerability::Relation::Vulnerability.def().rev())
+    }
 }
 
 impl Related<cvss3::Entity> for Entity {
@@ -66,16 +80,6 @@ impl Related<advisory_vulnerability::Entity> for Entity {
 impl Related<vulnerability_description::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Descriptions.def()
-    }
-}
-
-impl Related<advisory::Entity> for Entity {
-    fn to() -> RelationDef {
-        advisory_vulnerability::Relation::Advisory.def()
-    }
-
-    fn via() -> Option<RelationDef> {
-        Some(advisory_vulnerability::Relation::Vulnerability.def().rev())
     }
 }
