+++ modules/fetch/src/endpoints/vulnerability.rs
@@ -1,6 +1,8 @@
 use crate::endpoints::Error;
-use crate::model::vulnerability::Vulnerability;
+use crate::model::vulnerability::VulnerabilitySummary;
+use crate::service::FetchService;
 use actix_web::{get, web, HttpResponse, Responder};
+use trustify_common::db::Transactional;
 use trustify_common::model::{Paginated, PaginatedResults};
 use trustify_entity::vulnerability::Model;
 use trustify_module_search::model::SearchOptions;
@@ -18,20 +20,15 @@ use trustify_module_search::model::SearchOptions;
 )]
 #[get("")]
 pub async fn all(
+    state: web::Data<FetchService>,
     web::Query(search): web::Query<SearchOptions>,
     web::Query(paginated): web::Query<Paginated>,
 ) -> actix_web::Result<impl Responder> {
-    /*
-    let results = state
-        .vulnerabilities(search, paginated, ())
-        .await
-        .map_err(Error::System)?
-        .map(|e| Vulnerability::from(e.vulnerability));
-
-    Ok(HttpResponse::Ok().json(results))
-
-     */
-    Ok(HttpResponse::Ok().finish())
+    Ok(HttpResponse::Ok().json(
+        state
+            .fetch_vulnerabilities(search, paginated, Transactional::None)
+            .await?,
+    ))
 }
 
 #[utoipa::path(
@@ -41,30 +38,20 @@ pub async fn all(
         ("id", Path, description = "ID of the vulnerability")
     ),
     responses(
-        (status = 200, description = "Advisories related to the vulnerability"),
+        (status = 200, description = "Specified vulnerability"),
     ),
 )]
-#[get("/vulnerability/{id}/advisories")]
-pub async fn advisories() -> actix_web::Result<impl Responder> {
-    Ok(HttpResponse::Ok().finish())
-}
-
-#[utoipa::path(
-    context_path = "/api/v1/vulnerability",
-    tag = "vulnerability",
-    params(
-        ("id", Path, description = "ID of the vulnerability")
-    ),
-    responses(
-        (status = 200, description = "Affected packages"),
-    ),
-)]
-#[get("/vulnerability/{id}/affected/packages")]
-pub async fn affected_packages(
-    //state: web::Data<Graph>,
+#[get("/vulnerability/{id}")]
+pub async fn get(
+    state: web::Data<FetchService>,
     id: web::Path<String>,
 ) -> actix_web::Result<impl Responder> {
-    Ok(HttpResponse::Ok().finish())
+    let vuln = state.fetch_vulnerability(&id, Transactional::None).await?;
+    if let Some(vuln) = vuln {
+        Ok(HttpResponse::Ok().json(vuln))
+    } else {
+        Ok(HttpResponse::NotFound().finish())
+    }
 }
 
 #[utoipa::path(
